Feat/dao collab (#993)
* feat: Create Project Collaboration

* org search wip

* wip

* dao collab wip

* wip

* users list

* select with search

* wip, confirmation modal

* wip

* wip

* adding a suite of hotkeys to improve productivity (#930)

* adding a suite of hotkeys to improve productivity

* reverting graphql link back to staging

* testing vercel again

* fixing install bug

* fixing hotkey PR changes

* final draft revisions of hotkeys

* fix discussion error (#948)

* next draft

* remove console.logs

* adding guard classes

* changes to fix PR

* changes

* adding graphql url

* ready for review

* final change

* final revision of hotkey PR


* Guild token gating (#927)

* feat: import styles for token gatin

* feat: guild roles

* feat: guild token gating

* feat: guild token gating

* fix: ci

* fix: ci

* feat: change schema for guild

* feat: fix guild for role

* feat: re-use components

* feat: re-use components

* feat: re-use components

* feat: add tabs to the token-gating page

* feat: add tabs to the roles page

* fix minor ui bug

* fix: tabs spacing


* sidebar, approve, decline

* fix: Dropdown menu positioning using MUI (#969)

style: extract DropdownItem
style: remove unused Dropdown imports

* Revert "fix: Dropdown menu positioning using MUI (#969)" (#974)

This reverts commit 64e873923c13813ee82a2e42e3f8959d71a72a9f.

* Revert "Guild token gating (#927)" (#975)

This reverts commit 7829ef8528f170457bcfc0418b5d8df6b491fc58.

* Revert "adding a suite of hotkeys to improve productivity (#930)" (#976)

This reverts commit 49045652225a65c4035d38c0cd6785f4d07efff7.

* Feature/guild (#978)

* feat: import styles for token gatin

* feat: guild roles

* feat: guild token gating

* feat: guild token gating

* fix: ci

* fix: ci

* feat: change schema for guild

* feat: fix guild for role

* feat: re-use components

* feat: re-use components

* feat: re-use components

* feat: add tabs to the token-gating page

* feat: add tabs to the roles page

* fix minor ui bug

* fix: tabs spacing


* fix: Dropdown menu positioning using MUI (#977)

style: extract DropdownItem
style: remove unused Dropdown imports


* subset shortcuts

* Cypress setup & task tests (#907)

* Cypress setup & task tests

* Updates flow

* Test graphqlUri

* Revert latest

* Fix rebase linter issues

* Cleanup

* Updates

* Remove empty block on pod docs (#981)

Includes: Add button on docs on orgs and pods

* fix: Share link not working, refactor Share comp (#970)

* collaboration workspace

* subset hotkeys (#983)

* fix: rich text viewer for links

* feat: pagination on batch payment view (#984)

* fix: missing button text on Firefox (#989)

* fix org username

* modals flow

* task discord thread (#990)

* adding a suite of hotkeys to improve productivity (#930)

* adding a suite of hotkeys to improve productivity

* reverting graphql link back to staging

* testing vercel again

* fixing install bug

* fixing hotkey PR changes

* final draft revisions of hotkeys

* fix discussion error (#948)

* next draft

* remove console.logs

* adding guard classes

* changes to fix PR

* changes

* adding graphql url

* ready for review

* final change

* final revision of hotkey PR


* Guild token gating (#927)

* feat: import styles for token gatin

* feat: guild roles

* feat: guild token gating

* feat: guild token gating

* fix: ci

* fix: ci

* feat: change schema for guild

* feat: fix guild for role

* feat: re-use components

* feat: re-use components

* feat: re-use components

* feat: add tabs to the token-gating page

* feat: add tabs to the roles page

* fix minor ui bug

* fix: tabs spacing


* fix: Dropdown menu positioning using MUI (#969)

style: extract DropdownItem
style: remove unused Dropdown imports

* Revert "fix: Dropdown menu positioning using MUI (#969)" (#974)

This reverts commit 64e873923c13813ee82a2e42e3f8959d71a72a9f.

* Revert "Guild token gating (#927)" (#975)

This reverts commit 7829ef8528f170457bcfc0418b5d8df6b491fc58.

* Revert "adding a suite of hotkeys to improve productivity (#930)" (#976)

This reverts commit 49045652225a65c4035d38c0cd6785f4d07efff7.

* Feature/guild (#978)

* feat: import styles for token gatin

* feat: guild roles

* feat: guild token gating

* feat: guild token gating

* fix: ci

* fix: ci

* feat: change schema for guild

* feat: fix guild for role

* feat: re-use components

* feat: re-use components

* feat: re-use components

* feat: add tabs to the token-gating page

* feat: add tabs to the roles page

* fix minor ui bug

* fix: tabs spacing


* fix: Dropdown menu positioning using MUI (#977)

style: extract DropdownItem
style: remove unused Dropdown imports


* subset shortcuts

* Cypress setup & task tests (#907)

* Cypress setup & task tests

* Updates flow

* Test graphqlUri

* Revert latest

* Fix rebase linter issues

* Cleanup

* Updates

* Remove empty block on pod docs (#981)

Includes: Add button on docs on orgs and pods

* fix: Share link not working, refactor Share comp (#970)

* subset hotkeys (#983)

* fix: rich text viewer for links

* feat: pagination on batch payment view (#984)

* fix: missing button text on Firefox (#989)

* fix org username

* task discord thread


* wip

* feat: payment methods UI revamp (#991)

* chore: add icons for new payment methods ui

* feat: new payment method cta

* fix: activate and deactivate payment method

* feat: payment method empty state

* feat: add createdAt and deactivatedAt fields to PaymentMethodFragment

* chore: add colors for new payment methods ui

* chore: linter cleanup

* fix: payment methods sorting issue on activation / deactivation

* modals flow

* dao collab stars bg

* added skip for missing userid

+++ wondrous-app/components/Common/Filter/index.tsx
@@ -31,10 +31,11 @@ interface IFilterProps {
   onRemove?: ({}: TaskFilter) => void;
   selected?: any;
   key?: number;
+  withSearch?: boolean;
 }
 
 function Filter(props: IFilterProps) {
-  const { filterSchema = {}, onChange, currentIdx, schemaLength, onRemove, selected } = props;
+  const { filterSchema = {}, onChange, currentIdx, schemaLength, onRemove, selected, withSearch = false } = props;
   const { query, variables } = filterSchema;
   const [items, setItems] = useState(filterSchema?.items || []);
   const [open, setOpen] = useState(false);
@@ -87,7 +88,9 @@ function Filter(props: IFilterProps) {
   };
 
   const handleRemove = (filter, shouldCloseModal = true) => {
-    onRemove(filter);
+    if (onRemove) {
+      onRemove(filter);
+    }
     if (open) {
       setOpen(!shouldCloseModal);
     }
@@ -140,7 +143,7 @@ function Filter(props: IFilterProps) {
           <FilterValues>
             <Icon style={{ backgroundColor: '#0f0f0f', borderRadius: '6px' }} height="26" width="26" />
             {displayNames ? (
-              <InlineText>{`${filterSchema?.label}: ${displayNames}`}</InlineText>
+              <InlineText>{`${filterSchema?.label ? `${filterSchema?.label}: ` : ''} ${displayNames}`}</InlineText>
             ) : (
               filterSchema?.label || 'Filters'
             )}
